Search mechanism for subdirectories.

+++ ubiquitous_bash.sh
@@ -1771,10 +1771,29 @@ _setup_ssh_copyKey() {
 	sshKeyName="$1"
 	[[ "$2" != "" ]] && sshKeyLocalSubdirectory="$2"/
 	
-	chmod 600 "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName"
-	chmod 600 "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName".pub
-	cp -n "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName" "$sshLocalSSH"/"$sshKeyName"
-	cp -n "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName".pub "$sshLocalSSH"/"$sshKeyName".pub
+	if [[ -e "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName" ]]
+	then
+		chmod 600 "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName"
+		chmod 600 "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName".pub > /dev/null 2>&1
+		
+		cp -n "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName" "$sshLocalSSH"/"$sshKeyName"
+		cp -n "$scriptLocal"/ssh/"$sshKeyLocalSubdirectory""$sshKeyName".pub "$sshLocalSSH"/"$sshKeyName".pub > /dev/null 2>&1
+		
+		return 0
+	fi
+	
+	if [[ -e "$scriptLocal"/ssh/"$sshKeyName" ]]
+	then
+		chmod 600 "$scriptLocal"/ssh/"$sshKeyName"
+		chmod 600 "$scriptLocal"/ssh/"$sshKeyName".pub > /dev/null 2>&1
+		
+		cp -n "$scriptLocal"/ssh/"$sshKeyName" "$sshLocalSSH"/"$sshKeyName"
+		cp -n "$scriptLocal"/ssh/"$sshKeyName".pub "$sshLocalSSH"/"$sshKeyName".pub > /dev/null 2>&1
+		
+		return 0
+	fi
+	
+	return 1
 }
 
 #Overload with "ops".
